[ws-manager-mk2] Remove obsolete code
